Add support to base64 input for judger

+++ src/judger/judger.py
@@ -1,6 +1,7 @@
 import random
 import os
 import shutil
+import base64
 from .config import *
 from .compiler import Compiler
 from .runner import Runner
@@ -14,7 +15,11 @@ def copy_source_code(data, id):
         with open(path, "wb") as f:
             f.write(data['src'])
     elif isinstance(data['src'], str):
-        shutil.copy(data['src'], path)
+        if os.path.exists(data['src']):
+            shutil.copy(data['src'], path)
+        else:
+            with open(path, "w") as f:
+                f.write(base64.b64decode(data['src']).decode('utf-8'))
     else:
         raise TypeError("data['src'] should be bytes or str")
     return src_file
@@ -26,7 +31,11 @@ def copy_spj_code(data, id):
         with open(path, "wb") as f:
             f.write(data['spj_src'])
     elif isinstance(data['spj_src'], str):
-        shutil.copy(data['spj_src'], path)
+        if os.path.exists(data['spj_src']):
+            shutil.copy(data['spj_src'], path)
+        else:
+            with open(path, "w") as f:
+                f.write(base64.b64decode(data['spj_src']).decode('utf-8'))
     else:
         raise TypeError("data['spj_src'] should be bytes or str")
     return src_file
@@ -44,8 +53,14 @@ def copy_test_cases(data, id):
                 with open(os.path.join(DEFAULT_TMP_PATH, str(id), f'test_{idx}.out'), 'wb') as f:
                     f.write(test_output)
             elif isinstance(test_input, str):
-                shutil.copy(test_input, os.path.join(DEFAULT_TMP_PATH, str(id), f'test_{idx}.in'))
-                shutil.copy(test_output, os.path.join(DEFAULT_TMP_PATH, str(id), f'test_{idx}.out'))
+                if os.path.exists(test_input) and os.path.exists(test_output):
+                    shutil.copy(test_input, os.path.join(DEFAULT_TMP_PATH, str(id), f'test_{idx}.in'))
+                    shutil.copy(test_output, os.path.join(DEFAULT_TMP_PATH, str(id), f'test_{idx}.out'))
+                else:
+                    with open(os.path.join(DEFAULT_TMP_PATH, str(id), f'test_{idx}.in'), 'w') as f:
+                        f.write(base64.b64decode(test_input).decode('utf-8'))
+                    with open(os.path.join(DEFAULT_TMP_PATH, str(id), f'test_{idx}.out'), 'w') as f:
+                        f.write(base64.b64decode(test_output).decode('utf-8'))
             else:
                 raise TypeError("elements of data['test_case_input'] and data['test_case_output'] should be bytes or str")
     else:
@@ -104,15 +119,28 @@ def judge(data):
 
     #run
     logger.info("Running...")
-    for idx in range(len(data['test_case_input'])):
+    score = 0
+    results = {}
+    success = True
+    num_test_cases = len(data['test_case_input'])
+    if 'test_case_score' in data and isinstance(data['test_case_score'], list) and len(data['test_case_score']) == num_test_cases:
+        test_case_score = data['test_case_score']
+    else:
+        test_case_score = [0] * (num_test_cases - 1) + [100]
+    for idx in range(num_test_cases):
+        #TODO memory list
         runner = Runner(exe_path, data['language'], idx, id, data['max_time'], data['max_memory'], data.get('use_docker', False), data.get('use_spj', False), spj_exe_path)
         result = runner.run()
-        if not result['success']:
-            result['error_case'] = idx
-            return result
-
+        results[idx] = result
+        if result['success']:
+            score += test_case_score[idx]
+        else:
+            success = False
+        
+    results['score'] = score
+    results['success'] = success
     #clean
     shutil.rmtree(dir)
-    return result
+    return results
